# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Finance MVP API
  version: 0.1.0
  summary: Canonical OpenAPI contract for the mounted closed-MVP runtime surface.
  license:
    name: Proprietary/Internal
    url: https://example.invalid/proprietary-internal-license
  description: |
    This document is the source of truth for the mounted MVP wire routes, DTO field names,
    enum values, error codes, pagination envelopes, and auth requirements.

    MVP boundaries are intentionally narrow: financial data is manual-entry only,
    `sourceType` accepts only `manual`, and transfers are same-scope only. Safe
    screenshot OCR capture may create pending drafts from sanitized structured
    candidates, but raw image bytes and OCR text are rejected and confirmed
    transactions remain `sourceType=manual`. There is no real import, parser,
    file storage, bank API, automated ingestion, broker credential, external
    credential, card, IBAN, raw statement, or device-token endpoint.

    Privacy invariants:
    - personal data is owner-only;
    - shared data is visible only to active members of the same Household;
    - reports support `personal`, `shared_family_report`, and `combined_viewer_overview`;
    - every list, report, search, autocomplete, cache, count, and aggregation
      filters visible rows before any aggregation or pagination;
    - responses never expose hidden counts, filtered-out counts, hidden facets,
      or diagnostics that confirm inaccessible resources.
servers:
...
security:
  - pwaCookieAuth: []
    csrfHeader: []
  - androidBearerAuth: []
paths:
...
  /transactions:
    get:
      tags: [Transactions]
      operationId: listTransactions
      summary: List/search visible transactions.
      description: Transaction visibility inherits from visible account scope. Search, date, amount, and category filters apply after authz filtering.
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/AccountIdFilter'
        - $ref: '#/components/parameters/CategoryIdFilter'
        - $ref: '#/components/parameters/TransactionTypeFilter'
        - $ref: '#/components/parameters/HouseholdIdFilter'
        - $ref: '#/components/parameters/OwnershipTypeFilter'
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
        - $ref: '#/components/parameters/StatusFilter'
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: Visible transactions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionPageEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Transactions]
      operationId: createTransaction
      summary: Create manual transaction.
      description: |
        `sourceType` must be `manual`. For transfers, `accountId` and
        `counterpartyAccountId` must form one allowed same-scope transfer:
        `personal_same_owner` or `household_same_household`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionCreateRequest'
      responses:
        '201':
          description: Created transaction.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /transactions/autocomplete:
    get:
      tags: [Transactions]
      operationId: autocompleteTransactions
      summary: Autocomplete visible transactions.
      description: Minimal visible transaction labels only; no hidden matches or counts.
      parameters:
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Transaction suggestions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionAutocompleteListEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /transactions/{transactionId}:
    get:
      tags: [Transactions]
      operationId: getTransaction
      summary: Get visible transaction detail.
      parameters:
        - $ref: '#/components/parameters/TransactionId'
      responses:
        '200':
          description: Transaction detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
    patch:
      tags: [Transactions]
      operationId: updateTransaction
      summary: Update transaction mutable fields.
      description: Transfer updates must re-run full same-scope validation and cannot be partially applied.
      parameters:
        - $ref: '#/components/parameters/TransactionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionUpdateRequest'
      responses:
        '200':
          description: Updated transaction.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [Transactions]
      operationId: deleteTransaction
      summary: Soft delete transaction.
      parameters:
        - $ref: '#/components/parameters/TransactionId'
      responses:
        '204':
          description: Transaction soft-deleted or voided.
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /transactions/{transactionId}/restore:
    post:
      tags: [Transactions]
      operationId: restoreTransaction
      summary: Restore deleted transaction if allowed.
      parameters:
        - $ref: '#/components/parameters/TransactionId'
      responses:
        '200':
          description: Restored transaction.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
...
  /reports/transactions:
    get:
      tags: [Reports]
      operationId: getReportTransactions
      summary: Get report drill-down transactions.
      description: Every returned row must pass regular transaction list/detail predicates for the same actor.
      parameters:
        - $ref: '#/components/parameters/ReportMode'
        - $ref: '#/components/parameters/ReportHouseholdId'
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
        - $ref: '#/components/parameters/Timezone'
        - $ref: '#/components/parameters/AccountIds'
        - $ref: '#/components/parameters/CategoryIds'
        - $ref: '#/components/parameters/TransactionTypes'
        - $ref: '#/components/parameters/Currency'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: Report transaction drill-down.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportTransactionDrillDownEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    pwaCookieAuth:
      type: apiKey
      in: cookie
      name: __Host-finance_session
      description: HttpOnly Secure SameSite cookie used by the PWA.
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Required for PWA cookie-authenticated state-changing requests; bound to the authenticated session and rotated on login/logout/reset.
    androidBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Opaque bearer access token for Android; refresh token storage and rotation are platform secure-storage concerns.
  responses:
    Error:
      description: Canonical error envelope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  parameters:
...
    TransactionId:
      name: transactionId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/ResourceId' }
...
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
        minLength: 1
    Sort:
      name: sort
      in: query
      required: false
      schema:
        type: string
        minLength: 1
      description: Endpoint-specific allowlisted sort value.
    SearchQuery:
      name: q
      in: query
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 200
      description: Search runs only after visible-scope filtering and is not logged as raw payload.
    StartDate:
      name: startDate
      in: query
      required: true
      schema: { $ref: '#/components/schemas/DateOnly' }
    EndDate:
      name: endDate
      in: query
      required: true
      schema: { $ref: '#/components/schemas/DateOnly' }
    Timezone:
      name: timezone
      in: query
      required: true
      schema:
        type: string
        example: Europe/Moscow
    Currency:
      name: currency
      in: query
      required: false
      schema: { $ref: '#/components/schemas/CurrencyCode' }
...
    HouseholdIdFilter:
      name: householdId
      in: query
      required: false
      schema: { $ref: '#/components/schemas/ResourceId' }
      description: Optional scope narrowing only; inaccessible ids use neutral behavior.
    ReportHouseholdId:
      name: householdId
      in: query
      required: false
      schema: { $ref: '#/components/schemas/ResourceId' }
      description: Required for `shared_family_report` and `combined_viewer_overview`; omitted for `personal`.
    AccountIdFilter:
      name: accountId
      in: query
      required: false
      schema: { $ref: '#/components/schemas/ResourceId' }
    CategoryIdFilter:
      name: categoryId
      in: query
      required: false
      schema: { $ref: '#/components/schemas/ResourceId' }
    AccountIds:
      name: accountIds
      in: query
      required: false
      schema:
        type: array
        items: { $ref: '#/components/schemas/ResourceId' }
      style: form
      explode: false
      description: All supplied ids must be visible in the selected report mode or the request fails neutrally.
    CategoryIds:
      name: categoryIds
      in: query
      required: false
      schema:
        type: array
        items: { $ref: '#/components/schemas/ResourceId' }
      style: form
      explode: false
    TransactionTypes:
      name: transactionTypes
      in: query
      required: false
      schema:
        type: array
        items: { $ref: '#/components/schemas/TransactionType' }
      style: form
      explode: false
    ReportMode:
      name: reportMode
      in: query
      required: true
      schema: { $ref: '#/components/schemas/ReportMode' }
...
    OwnershipTypeFilter:
      name: ownershipType
      in: query
      required: false
      schema: { $ref: '#/components/schemas/OwnershipType' }
...
    TransactionTypeFilter:
      name: transactionType
      in: query
      required: false
      schema: { $ref: '#/components/schemas/TransactionType' }
    StatusFilter:
      name: status
      in: query
      required: false
      schema: { $ref: '#/components/schemas/RecordStatus' }
...
  schemas:
    ResourceId:
      type: string
      minLength: 1
      maxLength: 128
      pattern: '^[A-Za-z0-9_:-]+$'
      examples: [usr_123]
    DateTime:
      type: string
      format: date-time
      examples: ['2026-05-17T10:00:00Z']
    DateOnly:
      type: string
      format: date
      examples: ['2026-05-17']
...
    DecimalString:
      type: string
      pattern: '^-?[0-9]+(\.[0-9]{1,6})?$'
      examples: ['1000.00']
      description: Decimal-safe money string; never floating point.
    CurrencyCode:
      type: string
      pattern: '^[A-Z]{3}$'
      examples: [RUB]
    OwnershipType:
      type: string
      enum: [personal, shared]
...
    TransactionType:
      type: string
      enum: [income, expense, transfer, brokerage, asset_buy, asset_sell, interest, dividend, adjustment]
    ReportMode:
      type: string
      enum: [personal, shared_family_report, combined_viewer_overview]
      description: "`personal` includes only the current viewer's personal owner scope; household modes require householdId."
    TransferScope:
      type: string
      enum: [personal_same_owner, household_same_household]
      description: Public accepted transfer scopes are same-scope only. Unsupported cross-scope attempts are rejected with TRANSFER_SCOPE_NOT_SUPPORTED.
    TransferStatus:
      type: string
      enum: [posted, voided]
    SourceType:
      type: string
      enum: [manual]
      description: MVP accepts only manual source. Reserved post-MVP values are rejected by create/update flows.
      x-reserved-post-mvp-values: [file_import, bank_api, sms, push]
...
    RecordStatus:
      type: string
      enum: [active, archived, deleted]
...
    ErrorCode:
      type: string
      enum:
        - UNAUTHENTICATED
        - SESSION_EXPIRED
        - SESSION_REVOKED
        - RESOURCE_NOT_FOUND_OR_NOT_ACCESSIBLE
        - REFERENCED_RESOURCE_NOT_FOUND_OR_NOT_ACCESSIBLE
        - ACTION_NOT_ALLOWED
        - MEMBERSHIP_NOT_ACTIVE
        - VALIDATION_FAILED
        - INVALID_ENUM_VALUE
        - INVALID_DATE_RANGE
        - INVALID_CURRENCY
        - ACCOUNT_OWNERSHIP_IMMUTABLE
        - CATEGORY_SCOPE_MISMATCH
        - TRANSACTION_ACCOUNT_MISMATCH
        - ARCHIVED_RECORD_NOT_MUTABLE
        - CONFLICTING_UPDATE
        - ACCOUNT_CURRENCY_IMMUTABLE_AFTER_TRANSACTIONS
        - TRANSFER_SCOPE_NOT_SUPPORTED
        - TRANSFER_COUNTERPARTY_REQUIRED
        - TRANSFER_COUNTERPARTY_NOT_ACCESSIBLE
        - TOO_MANY_REQUESTS
        - RATE_LIMITED
        - LOGIN_RATE_LIMITED
        - RESET_RATE_LIMITED
        - INVITE_RATE_LIMITED
        - UPLOAD_TOO_LARGE
        - IMAGE_TOO_LARGE
        - UNSUPPORTED_MEDIA_TYPE
        - INVALID_IMAGE
        - OCR_DISABLED
        - OCR_ENGINE_UNAVAILABLE
        - OCR_TIMEOUT
        - INVITE_NOT_FOUND_OR_NOT_ACCESSIBLE
        - INVITE_EXPIRED
        - INVITE_ALREADY_USED
        - INVITE_REVOKED
        - HOUSEHOLD_MEMBER_LIMIT_REACHED
        - RESET_TOKEN_NOT_FOUND_OR_EXPIRED
        - RESET_TOKEN_ALREADY_USED
        - PASSWORD_POLICY_FAILED
        - IDEMPOTENCY_KEY_REUSED
        - UNSUPPORTED_ENTITY_TYPE
        - UNSUPPORTED_OPERATION
        - UNSUPPORTED_TRANSACTION_TYPE
        - PAYLOAD_REQUIRED
        - BASE_VERSION_REQUIRED
        - ENTITY_ALREADY_EXISTS
...
    ErrorDto:
      type: object
      additionalProperties: false
      required: [code, message, requestId]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: Safe generic user-facing message.
        requestId:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorDto'
    PageInfo:
      type: object
      additionalProperties: false
      required: [limit, hasMore]
      properties:
        limit:
          type: integer
          minimum: 1
          maximum: 100
        nextCursor:
          type: [string, 'null']
          description: Opaque cursor scoped to visible filtered rows only.
        hasMore:
          type: boolean
      description: Pagination metadata never includes totalCount or hidden counts.
...
    TransactionDto:
      type: object
      additionalProperties: false
      required: [id, transactionType, accountId, amount, currency, occurredAt, transactionDate, sourceType, createdByUserId, createdAt, updatedAt, version]
      properties:
        id:
          $ref: '#/components/schemas/ResourceId'
        transactionType:
          $ref: '#/components/schemas/TransactionType'
        accountId:
          $ref: '#/components/schemas/ResourceId'
        counterpartyAccountId:
          anyOf:
            - $ref: '#/components/schemas/ResourceId'
            - type: 'null'
        categoryId:
          anyOf:
            - $ref: '#/components/schemas/ResourceId'
            - type: 'null'
        amount:
          $ref: '#/components/schemas/DecimalString'
        currency:
          $ref: '#/components/schemas/CurrencyCode'
        occurredAt:
          $ref: '#/components/schemas/DateTime'
          description: Legacy compatible timestamp. When transactionDate is supplied by the client, the server stores this as stable UTC noon for that date.
        transactionDate:
          $ref: '#/components/schemas/DateOnly'
          description: Date-only operation date used for reporting and month boundaries.
        description:
          type: [string, 'null']
          maxLength: 500
          description: Shared transaction descriptions are visible to everyone who can read the shared transaction.
        sourceType:
          $ref: '#/components/schemas/SourceType'
        transferScope:
          anyOf:
            - $ref: '#/components/schemas/TransferScope'
            - type: 'null'
        transferStatus:
          anyOf:
            - $ref: '#/components/schemas/TransferStatus'
            - type: 'null'
        createdByUserId:
          $ref: '#/components/schemas/ResourceId'
        lastEditedByUserId:
          anyOf:
            - $ref: '#/components/schemas/ResourceId'
            - type: 'null'
        createdAt:
          $ref: '#/components/schemas/DateTime'
        updatedAt:
          $ref: '#/components/schemas/DateTime'
        deletedAt:
          anyOf:
            - $ref: '#/components/schemas/DateTime'
            - type: 'null'
        version:
          type: integer
          minimum: 1
    TransactionCreateRequest:
      type: object
      additionalProperties: false
      required: [transactionType, accountId, amount, currency, sourceType]
      properties:
        transactionType:
          $ref: '#/components/schemas/TransactionType'
        accountId:
          $ref: '#/components/schemas/ResourceId'
        counterpartyAccountId:
          $ref: '#/components/schemas/ResourceId'
          description: Required for transactionType=transfer.
        categoryId:
          $ref: '#/components/schemas/ResourceId'
          description: Not allowed for transfer in MVP.
        amount:
          $ref: '#/components/schemas/DecimalString'
        currency:
          $ref: '#/components/schemas/CurrencyCode'
        occurredAt:
          $ref: '#/components/schemas/DateTime'
          description: Legacy compatible timestamp. Required only when transactionDate is omitted.
        transactionDate:
          $ref: '#/components/schemas/DateOnly'
          description: Preferred date-only operation date. When both fields are supplied, transactionDate wins for reporting and occurredAt is normalized to UTC noon.
        description:
          type: string
          maxLength: 500
        sourceType:
          $ref: '#/components/schemas/SourceType'
      description: transactionDate is preferred; legacy clients may omit it only when occurredAt is supplied.
    TransactionUpdateRequest:
      type: object
      additionalProperties: false
      properties:
        transactionType:
          $ref: '#/components/schemas/TransactionType'
        accountId:
          $ref: '#/components/schemas/ResourceId'
        counterpartyAccountId:
          $ref: '#/components/schemas/ResourceId'
        categoryId:
          $ref: '#/components/schemas/ResourceId'
        amount:
          $ref: '#/components/schemas/DecimalString'
        currency:
          $ref: '#/components/schemas/CurrencyCode'
        occurredAt:
          $ref: '#/components/schemas/DateTime'
          description: Legacy compatible timestamp. Ignored for reporting date when transactionDate is also supplied.
        transactionDate:
          $ref: '#/components/schemas/DateOnly'
        description:
          type: string
          maxLength: 500
        sourceType:
          $ref: '#/components/schemas/SourceType'
        version:
          type: integer
          minimum: 1
      description: Client cannot set transferScope, transferStatus, actor fields, or balance effects.
    TransactionAutocompleteDto:
      type: object
      additionalProperties: false
      required: [id, transactionType, accountId, occurredAt, transactionDate, sourceType]
      properties:
        id:
          $ref: '#/components/schemas/ResourceId'
        transactionType:
          $ref: '#/components/schemas/TransactionType'
        accountId:
          $ref: '#/components/schemas/ResourceId'
        categoryId:
          anyOf:
            - $ref: '#/components/schemas/ResourceId'
            - type: 'null'
        occurredAt:
          $ref: '#/components/schemas/DateTime'
        transactionDate:
          $ref: '#/components/schemas/DateOnly'
        sourceType:
          $ref: '#/components/schemas/SourceType'
...
    ReportPeriodDto:
      type: object
      additionalProperties: false
      required: [startDate, endDate, timezone]
      properties:
        startDate:
          $ref: '#/components/schemas/DateOnly'
        endDate:
          $ref: '#/components/schemas/DateOnly'
        timezone:
          type: string
    ReportScopeDto:
      type: object
      additionalProperties: false
      required: [viewerUserId, reportMode, generatedAt]
      properties:
        viewerUserId:
          $ref: '#/components/schemas/ResourceId'
        householdId:
          anyOf:
            - $ref: '#/components/schemas/ResourceId'
            - type: 'null'
        reportMode:
          $ref: '#/components/schemas/ReportMode'
        includedAccountIds:
          type: array
          items:
            $ref: '#/components/schemas/ResourceId'
          description: Only visible account ids after report-mode and supplied-filter resolution.
        generatedAt:
          $ref: '#/components/schemas/DateTime'
...
    ReportTransactionDrillDownDto:
      type: object
      additionalProperties: false
      required: [scope, period, items, page]
      properties:
        scope:
          $ref: '#/components/schemas/ReportScopeDto'
        period:
          $ref: '#/components/schemas/ReportPeriodDto'
        items:
          type: array
          items:
            $ref: '#/components/schemas/TransactionDto'
        page:
          $ref: '#/components/schemas/PageInfo'
...
    TransactionEnvelope:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/TransactionDto'
    TransactionPageEnvelope:
      type: object
      additionalProperties: false
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TransactionDto'
        page:
          $ref: '#/components/schemas/PageInfo'
    TransactionAutocompleteListEnvelope:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TransactionAutocompleteDto'
...
    ReportTransactionDrillDownEnvelope:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/ReportTransactionDrillDownDto'